requirements-backend.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.3.0
sqlalchemy>=2.0.30
alembic>=1.13.1
psycopg[binary]>=3.1.19
redis>=5.0.4
celery>=5.4.0
pyjwt[crypto]>=2.8.0
